Fix some Eclipse warnings and adjust use of Generics, Comments, close()
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1691308 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/formula/functions/Countif.java
@@ -491,7 +491,7 @@ public final class Countif extends Fixed2ArgFunction {
 	 */
 	private static ValueEval evaluateCriteriaArg(ValueEval arg, int srcRowIndex, int srcColumnIndex) {
 		try {
-			return OperandResolver.getSingleValue(arg, srcRowIndex, (short)srcColumnIndex);
+			return OperandResolver.getSingleValue(arg, srcRowIndex, srcColumnIndex);
 		} catch (EvaluationException e) {
 			return e.getErrorEval();
 		}
